+++ surpyval/univariate/parametric/distributions/__init__.py
@@ -15,6 +15,7 @@ from .gamma import Gamma
 from .geometric import Geometric
 from .gumbel import Gumbel
 from .gumbel_lev import GumbelLEV
+from .hypoexponential import Hypoexponential
 from .logistic import Logistic
 from .loglogistic import LogLogistic
 from .lognormal import Galton, LogNormal